feat: add controller layer
