+++ internal/controller/rest/route_classes.go
@@ -1,4 +1,4 @@
-package main
+package rest
 
 import (
 	"context"
