+++ internal/controller/rest/route_users_test.go
@@ -1,4 +1,4 @@
-package main
+package rest
 
 import (
 	"github.com/PichuChen/go-bbs"
